requirements.txt
streamlit
wikipedia-api
pillow
pytesseract
requests